+++ proxy.ts
@@ -1,4 +1,4 @@
-export { auth as proxy } from "@/auth"
+export { auth as proxy } from "@/auth";
 export const config = {
   matcher: ["/((?!api|_next/static|_next/image|favicon.ico).*)"],
 }
